Provide packages for build/gfx-drm to use github.com/illumos/gfx-drm source
This change provide new packages for:

- driver/graphics/drm
- driver/graphics/agpgart

The userland library should be moved to an extra package which provide libdrm and could be part of omnios-extra. So the kernel module and the userland part are maintained in parallel but in different packages and repositories. Also libdrm could be a newer version and do not require the exact same version of the kernel module.

Because `illumos/gfx-drm` already provide make files to create packages these are used and transfered into the omnios repositories. Existing system tools are used to build the source for the kernel modules (no extra src/tools is build).

Versioning are based on the omnios version. The recommended option would be a fork of illumos/gfx-drm which than provide correct branches for each omnios version.