gcc compilers are not being built with optimisation
This problem was introduced during the recent work to move userland up to
gcc 9 which produces 64-bit binaries by default. As part of that, an explicit
"-m32" was added to the 32-bit CFLAGS and other environment variables since
the old configuration where this was just omitted for the 32-bit case now
resulted in 64-bit binaries being created.

gcc is built in a single configure/make pass and it generates both 32 and
64-bit components along the way. To make this work properly we need to
make sure that none of -m32 or -m64 are passed into environment variables
used by configure.

Unsetting CFLAGS had the unforseen effect of causing the compilers to be
build without optimisation, resulting in a significant performance penalty
when compiling code. For example, the illumos-omnios build time doubled
for several developers.